Create UIModal component that does not have proper hotkey Scope
