+++ packages/twenty-front/src/modules/auth/components/Modal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 
-import { Modal as UIModal } from '@/ui/layout/modal/components/Modal';
+import { UIModal } from '@/ui/layout/modal/components/UIModal';
 
 const StyledContent = styled(UIModal.Content)`
   align-items: center;
